% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_country_level_string.R
\name{build_country_level_string}
\alias{build_country_level_string}
\title{build_country_level_string
Create povcalnet query string for country level, survey year specification}
\usage{
build_country_level_string(country, poverty_line, survey_year,
  ppp = NULL, coverage_type = NULL)
}
\arguments{
\item{country}{character: Requested country. Use \href{https://www.iban.com/country-codes}{iso3 codes }}

\item{poverty_line}{numeric: Requested poverty lines in international dollars}

\item{survey_year}{numeric: Requested survey years}

\item{ppp}{numeric: Optional - Custom PPP exchange rates (Purchasing Power parity)}

\item{coverage_type}{character: Optional - Type of data coverage. Take one to the following values
rural; urban; national; special_cpi; aggregated_distribution, invalid_ppp}
}
\value{
character
}
\description{
build_country_level_string
Create povcalnet query string for country level, survey year specification
}
\examples{

build_country_level_string(country = c("ALB", "CHN"),
poverty_line = c(1.9, 2.0),
survey_year = c(2002, 2002))
}
