% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query_string.R
\name{build_query_string}
\alias{build_query_string}
\title{build_query_string
Create simplified povcalnet query string}
\usage{
build_query_string(country, poverty_line, year, aggregate = FALSE,
  fill_gaps = FALSE, coverage = "national", ppp = NULL,
  format = "json")
}
\arguments{
\item{country}{character: list of country iso3 code (accepts multiple) or `all`. Use \href{https://www.iban.com/country-codes}{iso3 codes }}

\item{poverty_line}{numeric: poverty line (in 2011 PPP-adjusted USD) to calculate poverty measures}

\item{year}{numeric:  list of years, or `all`, or `last`.}

\item{ppp}{numeric: Optional - Allows the selection of custom PPP (Purchasing Power Parity) exchange rates}

\item{display}{character: Specify whether the results should be displayed at
the country level, or as regional aggregates. Choose one of two options: "country_level",
"regional_aggregate"}
}
\value{
character
}
\description{
build_query_string
Create simplified povcalnet query string
}
\examples{

build_query_string(country = c("ALB", "CHN"),
poverty_line = 1.9,
year = c(2002, 2012))
}
