% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query_string_qp.R
\name{build_query_string_qp}
\alias{build_query_string_qp}
\title{build_query_string_qp
Create simplified povcalnet_qp query string}
\usage{
build_query_string_qp(
  country,
  popshare,
  year,
  aggregate = FALSE,
  fill_gaps = FALSE,
  coverage = "all",
  ppp = NULL,
  format = "json"
)
}
\arguments{
\item{country}{character: list of country iso3 code (accepts multiple) or `all`. Use \href{https://www.iban.com/country-codes}{iso3 codes }}

\item{popshare}{numeric: Percentage of the population who live below the poverty line}

\item{year}{numeric:  list of years, or `all`, or `last`.}

\item{aggregate}{logical: `TRUE` will return aggregate results,
`FALSE` country-level results.}

\item{fill_gaps}{logical: `TRUE` will interpolate / extrapolate values when
surveys are not available for a specific year.}

\item{coverage}{character: Can take one of four values: `all` (default), `national`, `urban`, `rural`.}

\item{ppp}{numeric: Optional - Allows the selection of custom PPP (Purchasing Power Parity) exchange rates}

\item{format}{character: Response format to be requested from the API:
`csv` or `json`}
}
\value{
character
}
\description{
build_query_string_qp
Create simplified povcalnet_qp query string
}
\examples{

build_query_string_qp(country = c("ALB", "CHN"),
popshare = .5,
year = c(2002, 2012))
}
