% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ref_year_string.R
\name{build_ref_year_string}
\alias{build_ref_year_string}
\title{build_ref_year_string
Create povcalnet query string for queries specifying reference years}
\usage{
build_ref_year_string(country, poverty_line, reference_year, ppp = NULL,
  display = "regional_aggregate")
}
\arguments{
\item{country}{character: Requested country. Use \href{https://www.iban.com/country-codes}{iso3 codes }}

\item{poverty_line}{numeric: Requested poverty line in international dollars}

\item{reference_year}{numeric: Requested survey years}

\item{ppp}{numeric: Optional - Custom PPP exchange rates (Purchasing Power parity)}

\item{display}{character: Specify whether the results should be displayed at
the country level, or as regional aggregates. Choose one of two options: "country_level",
"regional_aggregate"}
}
\value{
character
}
\description{
build_ref_year_string
Create povcalnet query string for queries specifying reference years
}
\examples{

build_ref_year_string(country = c("ALB", "CHN"),
poverty_line = 1.9,
reference_year = c(2002, 2012))
}
