% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_data_cl_json}
\alias{format_data_cl_json}
\title{format_data_cl_json}
\usage{
format_data_cl_json(x)
}
\arguments{
\item{x}{data.frame: API response}
}
\value{
data.frame
}
\description{
format_data_cl_json
}
