% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\title{get_countries}
\usage{
get_countries(region_code)
}
\arguments{
\item{region_code}{character: A region or income group code. Only accepts one
code}
}
\value{
a vector of country code
}
\description{
Helper function to quickly retrieve the list of iso3 country codes for a
specific World Bank region or income group
}
\examples{
get_countries("ECA")

}
