% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population.R
\name{get_population}
\alias{get_population}
\title{get_population}
\usage{
get_population(server = NULL)
}
\arguments{
\item{server}{character: Key for API root URL. For testing purposes only, should not be
changed for 99 percent of users.}
}
\value{
data.frame
}
\description{
get_population
}
\examples{
\donttest{
get_population()
}
}
