% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povcalnet_iterate.R
\name{povcalnet_iterate}
\alias{povcalnet_iterate}
\title{povcalnet_iterate
iterate to get monetary value of percentile selected (goal)}
\usage{
povcalnet_iterate(
  country = NULL,
  region = NULL,
  goal = 0.5,
  year = 2015,
  coverage = "national",
  pl = 1,
  tolerance = 5,
  ni = 40,
  delta = 3,
  aggregate = FALSE,
  fill_gaps = TRUE
)
}
\arguments{
\item{country}{character: ISO3 country code}

\item{region}{character: WB region code}

\item{goal}{numeric: desired percentile}

\item{year}{numeric: year}

\item{coverage}{character: aoiejw;}

\item{pl}{numeric: apoweij}

\item{tolerance}{numeric: apoweij}

\item{ni}{numeric: apoweij}

\item{delta}{numeric: apoweij}

\item{aggregate}{logical: apwoiejf}

\item{fill_gaps}{logical: `TRUE` will interpolate / extrapolate values when surveys are not available for a specific year.}
}
\value{
dataframe
}
\description{
povcalnet_iterate
iterate to get monetary value of percentile selected (goal)
}
\examples{
povcalnet_iterate(region= "WLD")
povcalnet_iterate(region= "WLD", goal = .4)
}
