% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povcalnet_qp.R
\name{povcalnet_qp}
\alias{povcalnet_qp}
\title{povcalnet_qp
Facilitate querying of a specific quantile of the population who live in poverty}
\usage{
povcalnet_qp(
  country = "all",
  quantile_poor = 0.2,
  year = "all",
  aggregate = FALSE,
  fill_gaps = FALSE,
  coverage = "all",
  ppp = NULL,
  server = NULL,
  format = "csv"
)
}
\arguments{
\item{country}{character: list of country iso3 code (accepts multiple) or
`all`. Use \code{\link{povcalnet_info}} for full list of countries.}

\item{quantile_poor}{numeric: Percentage of the population who live below the poverty line}

\item{year}{numeric:  list of years, or `all`.}

\item{aggregate}{logical: `TRUE` will return aggregate results,
`FALSE` country-level results.}

\item{fill_gaps}{logical: `TRUE` will interpolate / extrapolate values when
surveys are not available for a specific year.}

\item{coverage}{character: Can take one of three values: `national`, `urban`, `rural`}

\item{ppp}{numeric: Optional - Allows the selection of custom PPP (Purchasing Power Parity) exchange rates}

\item{server}{character: Key for API root URL. For testing purposes only, should not be
changed for 99 percent of users.}

\item{format}{character: Response format to be requested from the API:
`csv` or `json`}
}
\value{
data.frame
}
\description{
povcalnet_qp
Facilitate querying of a specific quantile of the population who live in poverty
}
\examples{
\donttest{
povcalnet_qp(country = "ALB")
}

}
