% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty_wb.R
\name{povcalnet_wb}
\alias{povcalnet_wb}
\title{povcalnet_wb
Returns the official WB regional estimates}
\usage{
povcalnet_wb(povline = 1.9, year = "all",
  url = "http://iresearch.worldbank.org", format = "csv")
}
\arguments{
\item{povline}{numeric: poverty line (in 2011 PPP-adjusted USD) to
calculate poverty measures}

\item{year}{numeric:  list of years, or `all`.}

\item{url}{character: API root URL. For testing purposes only, should not be
changed for 99 percent of users.}

\item{format}{character: Response format to be requested from the API:
`csv` or `json`}
}
\description{
povcalnet_wb
Returns the official WB regional estimates
}
