% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povcalnet_wb_qp.R
\name{povcalnet_wb_qp}
\alias{povcalnet_wb_qp}
\title{povcalnet_wb_qp
Returns the official WB regional estimates}
\usage{
povcalnet_wb_qp(popshare = 0.2, year = "all", server = NULL, format = "csv")
}
\arguments{
\item{popshare}{numeric: Percentage of the population who live below the poverty line}

\item{year}{numeric:  list of years, or `all`.}

\item{server}{character: Key for API root URL. For testing purposes only, should not be
changed for 99 percent of users.}

\item{format}{character: Response format to be requested from the API:
`csv` or `json`}
}
\value{
data.frame
}
\description{
povcalnet_wb_qp
Returns the official WB regional estimates
}
\examples{
\donttest{
povcalnet_wb_qp(year = 2015)
}
}
