% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_input}
\alias{sample_input}
\title{sample_input}
\format{A data frame with 5 rows and 5 variables:
\describe{
  \item{country}{iso3 country codes}
  \item{poverty_line}{Poverty line, in PPP}
  \item{year}{Survey year in YYYY format}
  \item{ppp}{PPP exchange rate}
  \item{coverage_type}{Survey coverage}
}}
\usage{
sample_input
}
\description{
A sample dataset to facilitate testing of `povcalnet_cl()`` function
}
\keyword{datasets}
