% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povcalnet_info.R
\name{povcal_info}
\alias{povcal_info}
\title{povcal_info}
\usage{
povcal_info(url = "http://iresearch.worldbank.org")
}
\arguments{
\item{url}{character: Path to file: This parameter should not be modified.}
}
\value{
data.frame
}
\description{
Non-memoised version of povcalnet_info()
}
