% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povcalnet_info.R
\name{povcalnet_info}
\alias{povcalnet_info}
\title{povcalnet_info()
Download updated information on available countries, regions, and surveys from the PovcalNet API}
\usage{
povcalnet_info(server = NULL)
}
\arguments{
\item{server}{character: Key for API root URL. For testing purposes only.
Should not be modified for 99 percent of users.}
}
\value{
data frame
}
\description{
povcalnet_info()
Download updated information on available countries, regions, and surveys from the PovcalNet API
}
\examples{
\donttest{
povcalnet_info()
}
}
