% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povcalnet_info.R
\name{povcalnet_info}
\alias{povcalnet_info}
\title{povcalnet_info()}
\usage{
povcalnet_info(url = "http://iresearch.worldbank.org")
}
\arguments{
\item{url}{character: This parameter should not be modified.}
}
\value{
data frame
}
\description{
Download updated information on available countries, regions, and surveys from the PovcalNet API
}
\examples{
\dontrun{
povcalnet_info()
}
}
